# order-persistor/docs/swagger.yaml
definitions:
  api.Error:
    properties:
      message:
        type: string
    type: object
  orders.Delivery:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      region:
        type: string
      zip:
        type: string
    required:
    - address
    - city
    - email
    - name
    - phone
    - region
    - zip
    type: object
  orders.Item:
    properties:
      brand:
        type: string
      chrt_id:
        type: integer
      name:
        type: string
      nm_id:
        type: integer
      price:
        type: number
      rid:
        type: string
      sale:
        type: number
      size:
        type: string
      status:
        type: integer
      total_price:
        type: number
      track_number:
        type: string
    type: object
  orders.Order:
    properties:
      customer_id:
        type: string
      date_created:
        type: string
      delivery:
        $ref: '#/definitions/orders.Delivery'
      delivery_service:
        type: string
      entry:
        type: string
      internal_signature:
        type: string
      items:
        items:
          $ref: '#/definitions/orders.Item'
        type: array
      locale:
        type: string
      oof_shard:
        type: string
      order_uid:
        type: string
      payment:
        $ref: '#/definitions/orders.Payment'
      shardkey:
        type: string
      sm_id:
        type: integer
      track_number:
        type: string
    required:
    - customer_id
    - date_created
    - delivery_service
    - entry
    - locale
    - oof_shard
    - order_uid
    - shardkey
    - sm_id
    - track_number
    type: object
  orders.Payment:
    properties:
      amount:
        type: number
      bank:
        type: string
      currency:
        type: string
      custom_fee:
        type: number
      delivery_cost:
        type: number
      goods_total:
        type: integer
      payment_dt:
        type: integer
      provider:
        type: string
      request_id:
        type: string
      transaction:
        type: string
    required:
    - amount
    - bank
    - currency
    - custom_fee
    - delivery_cost
    - goods_total
    - payment_dt
    - provider
    - transaction
    type: object
info:
  contact: {}
  title: Order-persistor API
  version: "1.0"
paths:
  /order/{id}:
    get:
      consumes:
      - application/json
      description: Returns the order object for the specified ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.Order'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get order by ID
      tags:
      - orders
swagger: "2.0"